genc

+++ lib/installers/install_nvm.sh
@@ -2,7 +2,9 @@
 
 function install_nvm() {
     curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.3/install.sh | bash
-    source "$HOME/.nvm/nvm.sh"
+    [[ -s "${HOME}/.nvm/nvm.sh" ]] || exit 1
+
+    source "${HOME}/.nvm/nvm.sh"
     nvm install --lts
 }
